@@ -31,12 +31,9 @@
     "test:previewnet:supplieradmin": "npx hardhat test --network previewnet test/supplieradmin.ts",
     "test:previewnet:wipeable": "npx hardhat test --network testnet test/wipeable.ts",
     "size": "npx hardhat size-contracts",
-    "build": "rimraf build && npx tsc -p tsconfig.json",
-    "build:types": "npx tsc -p tsconfig.json --emitDeclarationOnly",
-    "build:tsc": "rm -rf build && npx tsc -p tsconfig.json",
-    "build:watch": "npx tsc -w -p tsconfig.json",
-    "build:release": "npm run clean && tsc -p tsconfig.release.json",    
-    "lint": "npx solhint 'contracts/**/*.sol'",
+    "build": "rimraf build && npx tsc -p tsconfig.json", 
+    "lint:sol": "npx solhint 'contracts/**/*.sol'",
+    "lint:ts": "npx eslint . --ext .ts",
     "doc": "npx hardhat dodoc",
     "security": "docker run -t --rm -v /var/run/docker.sock:/var/run/docker.sock -v \"$(pwd)\":\"/home/ethsec/contracts\" -w \"/home/ethsec/contracts\" -u 0:0 trailofbits/eth-security-toolbox -c \"solc-select use 0.8.13  && slither ./contracts --solc-remaps @=node_modules/@\""
   },
@@ -47,34 +44,28 @@
     "@hashgraph/hethers": "^1.1.3",
     "@hashgraph/sdk": "2.17.1",
     "@nomicfoundation/hardhat-chai-matchers": "^1.0.2",
+    "@nomicfoundation/hardhat-network-helpers": "^1.0.6",
     "@nomicfoundation/hardhat-toolbox": "^1.0.2",
+    "@nomiclabs/hardhat-ethers": "^2.1.1",
+    "@nomiclabs/hardhat-etherscan": "^3.1.0",
     "@openzeppelin/contracts": "^4.7.2",
     "@openzeppelin/contracts-upgradeable": "^4.7.2",
     "@primitivefi/hardhat-dodoc": "^0.2.3",
-    "@nomicfoundation/hardhat-network-helpers": "^1.0.6",
-    "@nomiclabs/hardhat-ethers": "^2.1.1",
-    "@nomiclabs/hardhat-etherscan": "^3.1.0",
     "@typechain/ethers-v5": "^10.1.0",
     "@typechain/hardhat": "^6.1.3",
+    "@typescript-eslint/eslint-plugin": "^5.38.1",
+    "@typescript-eslint/parser": "^5.38.1",
+    "chai": "^4.3.6",
+    "eslint": "^8.24.0",
     "hardhat": "^2.10.1",
     "hardhat-abi-exporter": "^2.10.0",
     "hardhat-contract-sizer": "^2.6.1",
     "hardhat-gas-reporter": "^1.0.9",
-    "chai": "^4.3.6",
+    "ts-node": "^10.9.1",
     "typechain": "^8.1.0",
-	  "ts-node": "^10.9.1",
     "typescript": "latest"
   },
   "dependencies": {
-    "eslint": "^8.16.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-babel": "^5.3.1",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-mocha": "^10.0.4",
-    "eslint-plugin-n": "^15.2.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "ethlint": "^1.2.5",
     "solhint": "^3.3.7"
   }
 }
