@@ -1,6 +1,6 @@
-const { ContractId, AccountId } = require("@hashgraph/sdk");
-require("@hashgraph/hardhat-hethers");
-require("@hashgraph/sdk");
+const { ContractId, AccountId }  = require("@hashgraph/sdk");
+import "@hashgraph/hardhat-hethers";
+import "@hashgraph/sdk";
 
 import { expect } from "chai";
 import { deployContractsWithSDK, contractCall, getClient } from "../scripts/utils";
@@ -32,17 +32,17 @@ describe("Cash in tokens with maxTotalSupply", function() {
     expect(Number(resultTotalSupply[0])).to.equals(100000);
   });
   it("Cannot cash in 101 tokens to an account from a token with a 100 as maxTotalSupply", async function() {
-    let params = [AccountId.fromString(OPERATOR_ID!).toSolidityAddress(), 101000];        
+    const params = [AccountId.fromString(OPERATOR_ID!).toSolidityAddress(), 101000];        
     await expect(contractCall(ContractId.fromString(proxyAddress), 'mint', params, client, 400000, HederaERC20__factory.abi)).to.be.throw;
   });
   it("User without supplier_role cannot cash in tokens", async function() {
     const client2 = getClient();
     client2.setOperator(hreConfig.accounts[1].account, hreConfig.accounts[1].privateKey);      
-    let params = [AccountId.fromString(OPERATOR_ID!).toSolidityAddress(), 1000];        
+    const params = [AccountId.fromString(OPERATOR_ID!).toSolidityAddress(), 1000];        
     await expect(contractCall(ContractId.fromString(proxyAddress), 'mint', params, client2, 400000, HederaERC20__factory.abi)).to.be.throw;
   });
   it("Cannot cash in tokens to an account not associated to the token", async function() {
-    let params = [AccountId.fromString(OPERATOR_ID!).toSolidityAddress(), 1000];  
+    const params = [AccountId.fromString(OPERATOR_ID!).toSolidityAddress(), 1000];  
     await expect(contractCall(ContractId.fromString(proxyAddress), 'mint', params, client, 400000, HederaERC20__factory.abi)).to.be.throw;  
   });
 });
@@ -56,7 +56,7 @@ describe("Cash in tokens without maxTotalSupply", function() {
     proxyAddress = await deployContractsWithSDK("MIDAS", "MD", 3, 0, null, "Hedera Accelerator Stable Coin");    
   });
   it("Should cash in 10000 tokens to an account from a token without maxTotalSupply", async function() {
-    let params = [AccountId.fromString(OPERATOR_ID!).toSolidityAddress(), 10000000];  
+    const params = [AccountId.fromString(OPERATOR_ID!).toSolidityAddress(), 10000000];  
     await contractCall(ContractId.fromString(proxyAddress), 'mint', params, client, 400000, HederaERC20__factory.abi);
     const result = await contractCall(ContractId.fromString(proxyAddress), 'totalSupply', [], client, 60000, HederaERC20__factory.abi)  
     expect(Number(result[0])).to.equals(10000000);
