@@ -1,20 +1,19 @@
-const { ContractId, AccountId } = require("@hashgraph/sdk");
-require("@hashgraph/hardhat-hethers");
-require("@hashgraph/sdk");
+const { ContractId, AccountId }  = require("@hashgraph/sdk");
+import "@hashgraph/hardhat-hethers";
+import "@hashgraph/sdk";
 
 import { expect} from "chai";
 
 import { deployContractsWithSDK, contractCall, getClient } from "../scripts/utils";
 import { HederaERC20__factory } from "../typechain-types";
 
-const hre = require("hardhat"); 
+const hre = require("hardhat");
 const hreConfig = hre.network.config;
-const PAUSER_ROLE : string = '0x65d7a28e3265b37a6474929f336521b332c1681b933f6cb9f3376673440d862a';
+const PAUSER_ROLE  = '0x65d7a28e3265b37a6474929f336521b332c1681b933f6cb9f3376673440d862a';
 
 describe("Pausable", function() {
   let proxyAddress:any;
-  let client:any = getClient();;
-  let contractProxy: { name: (arg0: { gasLimit: number; }) => any; symbol: (arg0: { gasLimit: number; }) => any; decimals: (arg0: { gasLimit: number; }) => any; };
+  const client:any = getClient();
   let account:string;
   let privateKey:string;
 
@@ -22,9 +21,6 @@ describe("Pausable", function() {
     account = hreConfig.accounts[0].account;
     privateKey = hreConfig.accounts[0].privateKey;
     client.setOperator(account, privateKey);
-  });
-  beforeEach(async function () {
-    
   });
   it("Admin account can grant pauser role to an account", async function() {  
     proxyAddress = await deployContractsWithSDK("MIDAS", "MD", 3, 0, 1, "Hedera Accelerator Stable Coin");    
